+++ src/Telegram.Bot/Types/InlineQueryResults/Abstractions/IInputMessageContentResult.cs
@@ -5,6 +5,6 @@
         /// <summary>
         /// Content of the message to be sent
         /// </summary>
-        InputMessageContent InputMessageContent { get; set; }
+        InputMessageContentBase InputMessageContent { get; set; }
     }
 }