+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResultArticle.cs
@@ -19,7 +19,7 @@ namespace Telegram.Bot.Types.InlineQueryResults
 
         /// <inheritdoc />
         [JsonProperty(Required = Required.Always)]
-        public InputMessageContent InputMessageContent { get; set; }
+        public InputMessageContentBase InputMessageContent { get; set; }
 
         /// <summary>
         /// Optional. URL of the result.
@@ -66,7 +66,7 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// <param name="id">Unique identifier of this result</param>
         /// <param name="title">Title of the result</param>
         /// <param name="inputMessageContent">Content of the message to be sent</param>
-        public InlineQueryResultArticle(string id, string title, InputMessageContent inputMessageContent)
+        public InlineQueryResultArticle(string id, string title, InputMessageContentBase inputMessageContent)
             : this()
         {
             Id = id;