+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResultCachedAudio.cs
@@ -24,7 +24,7 @@ namespace Telegram.Bot.Types.InlineQueryResults
 
         /// <inheritdoc />
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public InputMessageContent InputMessageContent { get; set; }
+        public InputMessageContentBase InputMessageContent { get; set; }
 
         /// <summary>
         /// Initializes a new inline query result