+++ src/Telegram.Bot/Types/InlineQueryResults/InputContactMessageContent.cs
@@ -6,9 +6,8 @@ namespace Telegram.Bot.Types.InlineQueryResults
     /// <summary>
     /// Represents the content of a contact message to be sent as the result of an <see cref="InlineQuery"/>.
     /// </summary>
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn,
-                NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
-    public class InputContactMessageContent : InputMessageContent
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    public class InputContactMessageContent : InputMessageContentBase
     {
         /// <summary>
         /// Contact's phone number
@@ -25,7 +24,7 @@ namespace Telegram.Bot.Types.InlineQueryResults
         /// <summary>
         /// Optional. Contact's last name
         /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public string LastName { get; set; }
     }
 }