+++ test/Telegram.Bot.Tests.Integ/Update Messages/EditReplyMarkupTests.cs
@@ -68,7 +68,7 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
                 InlineKeyboardButton.WithCallbackData("Click here to change this button", data)
             });
 
-            InputMessageContent inputMessageContent = new InputTextMessageContent
+            InputMessageContentBase inputMessageContent = new InputTextMessageContent
             {
                 MessageText = "https://core.telegram.org/bots/api"
             };